+++ pkgs/standards/peagen/peagen/orm/works.py
@@ -1,0 +1,67 @@
+from __future__ import annotations
+
+from autoapi.v2.types import (
+    JSON,
+    Column,
+    PgUUID,
+    Integer,
+    ForeignKey,
+    relationship,
+    HookProvider,
+)
+from autoapi.v2.tables import Base
+from autoapi.v2.mixins import GUIDPk, Timestamped, StatusMixin
+
+from .tasks import Task
+
+
+class Work(Base, GUIDPk, Timestamped, StatusMixin, HookProvider):
+    __tablename__ = "works"
+    task_id = Column(PgUUID(as_uuid=True), ForeignKey("tasks.id"), nullable=False)
+    result = Column(JSON, nullable=True)
+    duration_s = Column(Integer)
+
+    task = relationship(Task, back_populates="works")
+
+    @classmethod
+    async def _post_update(cls, ctx):
+        from peagen.gateway import log, queue
+        from peagen.gateway._publish import _publish_task
+        from peagen.gateway.schedule_helpers import (
+            _finalize_parent_tasks,
+            _load_task,
+            _save_task,
+        )
+        from peagen.orm import Status
+
+        log.info("entering post_work_update")
+        wr = cls._SRead(**ctx["result"])
+        if not Status.is_terminal(wr.status):
+            return
+        task = await _load_task(queue, str(wr.task_id))
+        if task is None:
+            log.warning("terminal Work for unknown Task %s", wr.task_id)
+            return
+        updated = task.model_copy(
+            update={
+                "status": wr.status,
+                "result": wr.result or {},
+                "last_modified": wr.last_modified,
+            }
+        )
+        await _save_task(queue, updated)
+        await _publish_task(updated.model_dump(mode="json"))
+        await _finalize_parent_tasks(queue, str(wr.task_id))
+        log.info("Task %s closed via Work %s → %s", wr.task_id, wr.id, wr.status)
+
+    @classmethod
+    def __autoapi_register_hooks__(cls, api) -> None:
+        from autoapi.v2 import AutoAPI, Phase
+
+        cls._SRead = AutoAPI.get_schema(cls, "read")
+        api.register_hook(Phase.POST_COMMIT, model="Work", op="update")(
+            cls._post_update
+        )
+
+
+__all__ = ["Work"]
